requirements.txt
apptools==5.1.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
configobj==5.0.6
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
entrypoints==0.3
envisage==5.0.0
importlib-metadata==4.0.1
importlib-resources==5.1.2
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mayavi==4.7.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pyface==7.3.0
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
traitlets==5.0.5
traits==6.2.0
traitsui==7.1.1
typing-extensions==3.10.0.0
vtk==9.0.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
